new/backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
transformers
torch
dart-fss
yfinance
pydantic
sqlite3  # (일반적으로 표준 라이브러리)